Refactor MotorController command method to use int instead of int8_t and implement smoother acceleration/deceleration. Update begin method to set ESC frequency and attach with min/max pulse widths. Adjust motorTask commands to use microsecond values.

+++ esp32/microros_rplidar_c1/include/MotorController.h
@@ -10,7 +10,7 @@ public:
   bool begin();
 
   /// dir = +1 forward, -1 reverse, 0 stop/brake
-  void command(int8_t dir);
+  void command(int dir);
 
 private:
   int   _pwmPin;
